File system restructure for clarity/organization
