+++ BingoSync/CustomGoals/BingoGoal.cs
