+++ BingoSync/CustomGoals/GameMode.cs
