+++ BingoSync/CustomGoals/GameModesManager.cs
